Refactor global tick
